requirements.txt
absl-py==2.1.0
adagio==0.2.4
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.2
altair==5.0.1
ansi2html==1.8.0
antlr4-python3-runtime==4.11.1
argcomplete==3.1.1
async-timeout==4.0.2
autoviz==0.1.730
bidict==0.22.1
boto==2.49.0
boto3==1.28.25
botocore==1.31.25
cachetools==5.3.1
catboost==1.2
category-encoders==2.6.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.1.0
cloudpickle==2.2.1
colour==0.1.5
contourpy==1.0.7
crcmod==1.7
cycler==0.12.1
Cython==3.0.0
dacite==1.8.1
dash==2.11.1
dash-auth==2.0.0
dash-bootstrap-components==1.4.2
dash-core-components==2.0.0
dash-cytoscape==0.3.0
dash-html-components==2.0.0
dash-table==5.0.0
dash-testing-stub==0.0.2
dask==2023.7.0
databricks-cli==0.17.7
deepchecks==0.17.4
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.7
distributed==2023.7.0
docker==6.1.3
docopt==0.6.2
dtreeviz==2.2.2
einops==0.7.0
emoji==2.7.0
evidently==0.2.8
exceptiongroup==1.1.2
explainerdashboard==0.4.3
fairlearn==0.7.0
fastapi==0.101.0
fasteners==0.18
ffmpy==0.3.1
Flask==2.2.3
flask-simplelogin==0.1.2
Flask-WTF==1.1.1
fonttools==4.39.3
frozenlist==1.4.0
fs==2.4.16
fsspec==2023.6.0
fugue==0.8.6
fugue-sql-antlr==0.1.6
gcs-oauth2-boto-plugin==3.0
gevent==23.7.0
gitdb==4.0.10
GitPython==3.1.32
google-apitools==0.5.32
google-auth==2.22.0
google-reauth==0.1.1
gradio==3.40.1
gradio_client==0.4.0
graphviz==0.20.1
gsutil==5.25
h11==0.14.0
h5py==3.10.0
holoviews==1.14.9
htmlmin==0.1.12
httpcore==0.17.3
httplib2==0.20.4
httpx==0.24.1
huggingface-hub==0.16.4
hvplot==0.7.3
idna==3.6
ImageHash==4.3.1
imageio==2.34.0
importlib-metadata==5.2.0
importlib-resources==6.0.1
interpret==0.4.3
interpret-core==0.4.3
ipython==7.34.0
joblib==1.3.2
jsonpickle==3.0.1
jupyter-dash==0.4.2
kaleido==0.2.1
keras==3.1.1
kiwisolver==1.4.5
kmodes==0.12.2
lazy_loader==0.3
lightgbm==4.0.0
lime==0.2.0.1
linkify-it-py==2.0.2
llvmlite==0.40.1
m2cgen==0.10.0
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.8.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
ml-dtypes==0.3.2
mlflow==1.30.1
mlxtend==0.22.0
mne==1.2.3
monotonic==1.6
moto==4.1.14
msgpack==1.0.5
multidict==6.0.4
multimethod==1.9.1
multiprocess==0.70.15
namex==0.0.7
networkx==3.2.1
numba==0.57.1
numpy==1.26.2
oauth2client==4.1.3
oauthlib==3.2.2
optree==0.11.0
orjson==3.9.4
outcome==1.2.0
oyaml==1.0
packaging==21.3
pandas==1.5.3
pandas-dq==1.28
pandas-profiling==3.6.6
partd==1.4.0
percy==2.0.2
phik==0.12.3
Pillow==9.3.0
platformdirs==3.2.0
plotly==5.16.0
plotly-resampler==0.9.1
pmdarima==2.0.3
pooch==1.7.0
prometheus-flask-exporter==0.22.4
protobuf==4.24.0
psutil==5.9.5
pyamg==5.0.1
pyarrow==12.0.1
pycaret==3.0.4
pycparser==2.21
pydub==0.25.1
Pygments==2.17.2
pynndescent==0.5.10
PyNomaly==0.3.3
pyod==1.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.6
python-utils==3.7.0
pytz==2022.7.1
pyu2f==0.1.5
qpd==0.4.4
querystring-parser==1.2.4
requests==2.31.0
responses==0.23.1
retry-decorator==1.1.1
retrying==1.3.4
rich==13.7.1
rsa==4.7.2
s3transfer==0.6.1
SALib==1.4.7
schemdraw==0.15
scikit-base==0.5.0
scikit-image==0.22.0
scikit-learn==1.3.2
scikit-plot==0.3.7
scipy==1.10.1
selenium==4.2.0
semantic-version==2.10.0
shap==0.42.1
shortuuid==1.0.11
six==1.16.0
skope-rules==1.0.1
sktime==0.21.0
slicer==0.0.7
smmap==5.0.0
soundfile==0.12.1
sqlglot==17.12.0
sqlparse==0.4.4
starlette==0.27.0
statsforecast==1.5.0
stumpy==1.11.1
tangled-up-in-unicode==0.2.0
tbats==1.1.3
tblib==2.0.0
textblob==0.17.1
threadpoolctl==3.1.0
tifffile==2024.2.12
torch ==1.11.0
torchvision==0.12.0
tornado==6.3.2
tqdm==4.65.0
trace-updater==0.0.9.1
treeinterpreter==0.2.3
triad==0.9.1
trio==0.22.2
trio-websocket==0.10.3
tsdownsample==0.1.2
tsfresh==0.20.1
typeguard==2.13.3
types-PyYAML==6.0.12.11
typing_extensions==4.10.0
tzdata==2023.3
uc-micro-py==1.0.2
umap-learn==0.5.3
urllib3==1.26.7
urllib3-secure-extra==0.1.0
uvicorn==0.23.2
visions==0.7.5
waitress==2.1.2
websockets==11.0.3
wfdb==4.1.2
wordcloud==1.9.2
wsproto==1.2.0
WTForms==3.0.1
xgboost==1.7.6
xlrd==2.0.1
xmltodict==0.13.0
xxhash==3.3.0
yarl==1.9.2
ydata-profiling==4.5.1
yellowbrick==1.5
zict==3.0.0
zipp==3.16.2
zope.event==5.0
